DRY out request_build() return object
I used to retain 'path' and 'query'  as separate components, even though they have been incorporated into the 'url' at this point, because it was useful to see them, esp. during development.

But it seems better in the long-run to remove them. Presence of the same info in two places, prior to calling request_make(), makes it possible for someone to do weird things. 'path' and 'query' should be regarded as read-only but I have no way to enforce that.