Repair tokens stored to path unequal to their hash
Closes #109, closes https://github.com/tidyverse/googledrive/issues/274

Deals with cache loading failure in the presence of oauth tokens cached on R 3.5, but then accessed on R 3.6. Relates to the change to the default serialization version (2 -> 3).

Also dropped logic around checking for duplicates because it's extremely difficult to have 2 or more files (tokens) with the same name. Seems unlikely to affect our cache.