Minor upgrades to Discovery Document processing
Overall goal is to make it easier to identify and use an existing Discovery Document, as opposed to always downloading the current version.

I'm seeing a lot of nuisance diffs with Sheets v4 JSON, i.e. the JSON changes but is semantically equivalent to previous version. I want it to be easier to be more intentional about when I re-download.