Simplify logic around Retry-After based on better understanding of case insensitive headers
It is now clear that https://github.com/r-lib/httr/issues/611 is actually a bug around `quiet = TRUE`. httr's insensitive class should always be used to construct the header list, which is indeed working correctly in httr itself. But my original tests failed to do this and I was confused for a while about what the problem was.