Start to use the new client type ("gargle_oauth_client") for installed clients
I originally did this because I wanted httr's default logic for the redirect URI. But we already set the redirect URI explicitly via oob_value for conventional or pseudo-oob.

So that only leaves the "temporary web server" flow and it makes more sense to directly inject httr's default, which is httr::oauth_callback().