Complete the switch from "app" to "client" (#257)
* "Application" instead of "app"

* AuthState

* Gargle2.0 and oauth cache

* credentials_user_oauth2()

* Don't cache this weird test token

* Use withCallingHandlers()

The original approach, with only tryCatch(), meant that the deprecation warnings would cause an exit, which is contrary to the design of token_fetch().

* Switch to deprecate_soft()

I've done a bit of work with wrapper packages now, prior to updating them, and I think these should be soft deprecations at first.

* Make the redirect_uris more plausible for the example web client

* Switch this one back to deprecate_warn()

* Switch this to deprecate_warn() also

No "normal" user should be calling Gargle2.0$new(). They should be using gargle2.0_token(). So I'm happy to be a bit aggressive with the warning if someone does Gargle2.0$new(app).

* Tweak comments

* Work on this deprecation message

I am, however, beginning to think that making `app` an active field is overall a negative. I'm probably going to back out of that. But I'd like to capture this state in case I come back to it.

* Improve these messages

Because the R6-ness, I think the lifecycle defaults can't be relied upon.

* More conventional arrangement of roxygen comment

* Update the vignette about gargle auth design

* WIP

* Add a vignette for maintainers

* Don't use an active field after all

I don't think the pros outweight the cons.

Pros:
* Generates a deprecation message, so actively discourages use.
* Feels elegant.

Cons:
* Somehow circumvents lifecycle::deprecate_soft(), so the message shows up even for users who haven't tried to directly access $app.
* Makes the class more complicated.
* Makes the `app` treatment in AuthState and Gargle2.0 less parallel.

* Don't label this as experimental any more

* Re-document()

* Add NEWS bullet