Add support for viewer-based credentials on Posit Connect.
This commit wires up a new credential provider for Connect's
viewer-based credentials feature [0]. Most of the actual work is
outsourced to a new shared package, `connectcreds` [1].

Viewer credentials are inherently tied to a given Shiny session, which
is at odds with `gargle`'s existing view that a single credential is
active for a given R process -- so we need to unwind this assumption.

In order to support storing and retrieving "session" credentials from
the existing `AuthState` object -- while preserving backward- and
forward-compatibility with existing packages -- I have modified its API
so that setters and getters are now aware of the existince of session
credentials. (There are comprehensive unit tests that explain the
details.)

Existing packages need to be updated to use this getter, if they aren't
already (most are not), but in the meantime they will continue to work
-- though they won't be able to use viewer-based credentials. This seems
like a reasonable tradeoff, and allows us to say e.g. "upgrade
`bigrquery` if you want to use it with Connect's viewer-based
credentials".

Unit tests are included.

[0]: https://docs.posit.co/connect/user/oauth-integrations/
[1]: https://github.com/posit-dev/connectcreds/

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>